requirements.txt
langchain_openai
langchain
python-dotenv
streamlit